# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VPS Billing Platform API
  version: 2.0.0
  license: {name: Proprietary}
servers:
  - url: /api/v1
security: [{userSession: []}]
paths:
...
  /instances/{id}:
    get:
      summary: Get an instance
      operationId: getInstance
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Instance}
        "404": {description: Not found}
  /instances/{id}/networks:
    get:
      summary: List instance network addresses owned by the current user
      operationId: listInstanceNetworks
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses:
        "200": {description: Networks}
        "404": {$ref: "#/components/responses/NotFound"}
  /instances/{id}/traffic:
    get:
      summary: List recent traffic samples for an owned instance
      operationId: listInstanceTraffic
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses:
        "200": {description: Traffic samples}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /notifications/{id}/read:
    put:
      summary: Mark an owned notification read
      operationId: markNotificationRead
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "200": {description: Marked read}
        "404": {$ref: "#/components/responses/NotFound"}
  /tickets:
    get:
      summary: List current user support tickets
      operationId: listTickets
      responses:
        "200": {description: Tickets}
    post:
      summary: Create a support ticket with its first message
      operationId: createTicket
      parameters: [{$ref: "#/components/parameters/CSRFToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, priority, message]
              additionalProperties: false
              properties:
                subject: {type: string, minLength: 3, maxLength: 255}
                priority: {type: string, enum: [low, normal, high]}
                message: {type: string, minLength: 1, maxLength: 10000}
      responses:
        "201": {description: Ticket created}
        "422": {$ref: "#/components/responses/ValidationFailure"}
  /tickets/{id}:
    get:
      summary: Get an owned support ticket and messages
      operationId: getTicket
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses:
        "200": {description: Ticket detail}
        "404": {$ref: "#/components/responses/NotFound"}
  /tickets/{id}/messages:
    post:
      summary: Add a user message to an owned support ticket
      operationId: addTicketMessage
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              additionalProperties: false
              properties: {message: {type: string, minLength: 1, maxLength: 10000}}
      responses:
        "201": {description: Message created}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /instances/{id}/usage:
    get: {summary: Get current usage allowance and estimated charge, parameters: [{$ref: "#/components/parameters/Id"}], responses: {"200": {description: Usage summary}, "404": {$ref: "#/components/responses/NotFound"}}}
  /instances/{id}/port-forwards:
    get: {summary: List owned NAT port-forward mappings, parameters: [{$ref: "#/components/parameters/Id"}], responses: {"200": {description: Port-forward mappings}}}
    post: {summary: Queue a NAT port-forward addition, parameters: [{$ref: "#/components/parameters/Id"}, {$ref: "#/components/parameters/CSRFToken"}], responses: {"202": {description: Accepted, content: {application/json: {schema: {$ref: "#/components/schemas/OperationAccepted"}}}}, "409": {$ref: "#/components/responses/Conflict"}}}
...
components:
  securitySchemes:
    userSession: {type: apiKey, in: cookie, name: vps_user_session}
    adminSession: {type: apiKey, in: cookie, name: vps_admin_session}
...
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
...
  responses:
...
    Conflict: {description: Resource conflict, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    ValidationFailure: {description: Validation failure, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, const: false}
        request_id: {type: string}
        error:
          type: object
          required: [code, message_key]
          properties:
            code: {type: string}
            message_key: {type: string}
            details: {type: object, additionalProperties: true}
...
    OperationAccepted:
      type: object
      required: [success, data, request_id]
      properties:
        success: {type: boolean, const: true}
        request_id: {type: string}
        data:
          type: object
          required: [operation_id, status]
          properties:
            operation_id: {type: string, format: uuid}
            status: {type: string, enum: [queued]}
...